feat: apply validation

+++ src/main/java/dev/sv/cloud_file_storage/service/ResourceServiceImpl.java
@@ -2,9 +2,7 @@ package dev.sv.cloud_file_storage.service;
 
 import dev.sv.cloud_file_storage.dto.ResourceDto;
 import dev.sv.cloud_file_storage.entity.User;
-import dev.sv.cloud_file_storage.exception.InvalidFilenameException;
-import dev.sv.cloud_file_storage.exception.InvalidOperationException;
-import dev.sv.cloud_file_storage.exception.ResourceAlreadyExistsException;
+import dev.sv.cloud_file_storage.exception.*;
 import dev.sv.cloud_file_storage.mapper.ResourceMapper;
 import dev.sv.cloud_file_storage.utils.Path;
 import dev.sv.cloud_file_storage.utils.PathUtils;
@@ -34,9 +32,13 @@ import static dev.sv.cloud_file_storage.utils.PathUtils.isDirectory;
 @RequiredArgsConstructor
 public class ResourceServiceImpl implements ResourceService {
 
+    private static final String RESOURCE_ALREADY_EXISTS = "Resource already exists";
+    private static final String RESOURCE_NOT_FOUND = "Resource under path '%s' doesn't exist";
+    private static final String DIRECTORY_ALREADY_EXISTS = "Directory already exists";
+    private static final String DIRECTORY_NOT_FOUND = "Directory under path '%s' doesn't exist";
+    private static final String MISSING_DIRECTORY = "Missing directory! Check out whether directories by path '%s' exist";
+    private static final String INVALID_FILE_NAME = "Invalid file name";
     private static final String USER_PREFIX = "user-%s-files/";
-    public static final String RESOURCE_ALREADY_EXISTS = "Resource already exists";
-    public static final String INVALID_FILE_NAME = "Invalid file name";
 
     private final MinioService minioService;
 
@@ -45,20 +47,34 @@ public class ResourceServiceImpl implements ResourceService {
     @Override
     public ResourceDto getResource(String path, Long userId) {
         Path total = new Path(path, userId);
-        StatObjectResponse sor = minioService.statObject(total.toString());
 
+        if (!minioService.objectExists(total.getAbsolutePath())) {
+            throw new ResourceNotFoundException(RESOURCE_NOT_FOUND.formatted(total.getNormalPath()));
+        }
+
+        StatObjectResponse sor = minioService.statObject(total.getAbsolutePath());
         return resourceMapper.map(sor.size(), new Path(sor.object(), userId));
     }
 
     @Override
-    public void deleteResource(String path, Long userId)  {
+    public void deleteResource(String path, Long userId) {
         Path total = new Path(path, userId);
+
+        if (!minioService.objectExists(total.getAbsolutePath())) {
+            throw new ResourceNotFoundException(RESOURCE_NOT_FOUND.formatted(total.getNormalPath()));
+        }
+
         minioService.removeObject(total.getAbsolutePath());
     }
 
     @Override
     public void downloadResource(String path, Long userId, HttpServletResponse response) {
         Path total = new Path(path, userId);
+
+        if (!minioService.objectExists(total.getAbsolutePath())) {
+            throw new ResourceNotFoundException(RESOURCE_NOT_FOUND.formatted(total.getNormalPath()));
+        }
+
         try {
             if (!total.isDirectory()) {
                 downloadFile(path, userId, response);
@@ -72,6 +88,135 @@ public class ResourceServiceImpl implements ResourceService {
                  InternalException e) {
             throw new RuntimeException(e);
         }
+    }
+
+    @Override
+    public ResourceDto moveResource(String from, String to, Long userId) {
+        Path origin = new Path(from, userId);
+        Path target = new Path(to, userId);
+
+        if (!minioService.objectExists(origin.getAbsolutePath())) {
+            throw new ResourceNotFoundException(RESOURCE_NOT_FOUND.formatted(origin.getNormalPath()));
+        }
+        if (minioService.objectExists(target.getAbsolutePath())) {
+            throw new ResourceAlreadyExistsException(RESOURCE_ALREADY_EXISTS);
+        }
+
+        if (origin.getPathWithoutPrefixAndFile().equals(target.getPathWithoutPrefixAndFile())) {
+            rename(origin, target);
+        } else {
+            move(origin, target);
+        }
+
+        StatObjectResponse sor = minioService.statObject(target.getAbsolutePath());
+        return resourceMapper.map(sor.size(), new Path(sor.object(), userId));
+    }
+
+    @Override
+    public List<ResourceDto> searchResource(String query, Long userId) {
+        Iterable<Result<Item>> results = minioService.listObjects(PathUtils.appendRootPrefix("", userId), true);
+        List<ResourceDto> resources = new ArrayList<>();
+        for (Result<Item> result : results) {
+            try {
+                Item item = result.get();
+                String object = item.objectName();
+                if (object.contains(query)) {
+                    resources.add(resourceMapper.map(item.size(), new Path(object, userId)));
+                }
+            } catch (ErrorResponseException | InsufficientDataException | InternalException | InvalidKeyException |
+                     InvalidResponseException | IOException | NoSuchAlgorithmException | ServerException |
+                     XmlParserException e) {
+                throw new RuntimeException(e);
+            }
+        }
+        return resources;
+    }
+
+    @Override
+    public List<ResourceDto> uploadResource(String path, MultipartFile[] files, Long userId) {
+        Path total = new Path(path, userId);
+        List<String> paths = new ArrayList<>();
+        for (MultipartFile file : files) {
+            String fileName = file.getOriginalFilename();
+            if (fileName == null) {
+                throw new InvalidFileNameException(INVALID_FILE_NAME);
+            }
+            if (fileName.contains("/")) {
+                createRecursiveDirectories(userId, fileName);
+            }
+            if (minioService.objectExists(total.getAbsolutePath() + fileName)) {
+                throw new ResourceAlreadyExistsException(RESOURCE_ALREADY_EXISTS);
+            }
+
+            paths.add(total.getAbsolutePath() + fileName);
+            minioService.putObject(total.getAbsolutePath() + fileName, file);
+        }
+
+        List<ResourceDto> resources = new ArrayList<>();
+        for (String p : paths) {
+            StatObjectResponse sor = minioService.statObject(p);
+            resources.add(resourceMapper.map(sor.size(), new Path(sor.object(), userId)));
+        }
+        return resources;
+    }
+
+    @Override
+    public ResourceDto createDirectory(String path, Long userId) {
+        path = path.endsWith("/") ? path : path + "/";
+        Path total = new Path(path, userId);
+
+        if (minioService.objectExists(total.getAbsolutePath())) {
+            throw new ResourceAlreadyExistsException(DIRECTORY_ALREADY_EXISTS);
+        }
+
+        ObjectWriteResponse owr = minioService.putEmptyObject(total.getAbsolutePath());
+        return resourceMapper.map(null, new Path(owr.object(), userId));
+    }
+
+    @Override
+    public List<ResourceDto> getDirectory(String path, Long userId) {
+        Path total = new Path(path, userId);
+
+        if (!minioService.objectExists(total.getAbsolutePath())) {
+            throw new ResourceNotFoundException(DIRECTORY_NOT_FOUND.formatted(total.getNormalPath()));
+        }
+
+        Iterable<Result<Item>> results = minioService.listObjects(total.getAbsolutePath(), true);
+        List<ResourceDto> resources = new ArrayList<>();
+        for (Result<Item> result : results) {
+            try {
+                Item item = result.get();
+                resources.add(resourceMapper.map(item.size(), new Path(item.objectName(), userId)));
+            } catch (ErrorResponseException | InsufficientDataException | InternalException | InvalidKeyException |
+                     InvalidResponseException | IOException | NoSuchAlgorithmException | ServerException |
+                     XmlParserException e) {
+                throw new RuntimeException(e);
+            }
+        }
+        return resources;
+    }
+
+    @Override
+    public void createUserDirectory(User user) {
+        minioService.putEmptyObject(USER_PREFIX.formatted(user.getId()));
+    }
+
+    private void rename(Path source, Path target) {
+        if (source.getFileName().equals(target.getFileName())) {
+            throw new InvalidOperationException("Target resource should have different name");
+        }
+
+        minioService.copyObject(target.getAbsolutePath(), source.getAbsolutePath());
+        minioService.removeObject(source.getAbsolutePath());
+    }
+
+    private void move(Path source, Path target) {
+        if (!source.getFileName().equals(target.getFileName())) {
+            throw new InvalidOperationException("Target resource should have same name");
+        }
+
+        minioService.copyObject(target.getAbsolutePath(), source.getAbsolutePath());
+        minioService.removeObject(source.getAbsolutePath());
     }
 
     private void downloadFile(String path, Long userId, HttpServletResponse response) throws IOException {
@@ -114,128 +259,15 @@ public class ResourceServiceImpl implements ResourceService {
         }
     }
 
-    @Override
-    public ResourceDto moveResource(String from, String to, Long userId) {
-        Path origin = new Path(from, userId);
-        Path target = new Path(to, userId);
-
-        if (origin.getPathWithoutPrefixAndFile().equals(target.getPathWithoutPrefixAndFile())) {
-            rename(origin, target);
-        } else {
-            move(origin, target);
-        }
-
-        StatObjectResponse sor = minioService.statObject(target.getAbsolutePath());
-        return resourceMapper.map(sor.size(), new Path(sor.object(), userId));
-    }
-
-    private void rename(Path source, Path target) {
-        if (source.getFileName().equals(target.getFileName())) {
-            throw new InvalidOperationException("Target resource should have different name");
-        }
-
-        minioService.copyObject(target.getAbsolutePath(), source.getAbsolutePath());
-        minioService.removeObject(source.getAbsolutePath());
-    }
-
-    private void move(Path source, Path target) {
-        if (!source.getFileName().equals(target.getFileName())) {
-            throw new InvalidOperationException("Target resource should have same name");
-        }
-
-        minioService.copyObject(target.getAbsolutePath(), source.getAbsolutePath());
-        minioService.removeObject(source.getAbsolutePath());
-    }
-
-    @Override
-    public List<ResourceDto> searchResource(String query, Long userId) {
-        Iterable<Result<Item>> results = minioService.listObjects(PathUtils.appendRootPrefix("", userId), true);
-        List<ResourceDto> resources = new ArrayList<>();
-        for (Result<Item> result : results) {
-            try {
-                Item item = result.get();
-                String object = item.objectName();
-                if (object.contains(query)) {
-                    resources.add(resourceMapper.map(item.size(), new Path(object, userId)));
-                }
-            } catch (ErrorResponseException | InsufficientDataException | InternalException | InvalidKeyException |
-                     InvalidResponseException | IOException | NoSuchAlgorithmException | ServerException |
-                     XmlParserException e) {
-                throw new RuntimeException(e);
-            }
-        }
-        return resources;
-    }
-
-    @Override
-    public List<ResourceDto> uploadResource(String path, MultipartFile[] files, Long userId) {
-        Path total = new Path(path, userId);
-        List<String> paths = new ArrayList<>();
-        for (MultipartFile file : files) {
-            String fileName = file.getOriginalFilename();
-            if (fileName == null) {
-                throw new InvalidFilenameException(INVALID_FILE_NAME);
-            }
-            if (fileName.contains("/")) {
-                createRecursiveDirectories(userId, fileName);
-            }
-            if (minioService.objectExists(total.getAbsolutePath() + fileName)) {
-                throw new ResourceAlreadyExistsException(RESOURCE_ALREADY_EXISTS);
-            }
-
-            paths.add(total.getAbsolutePath() + fileName);
-            minioService.putObject(total.getAbsolutePath() + fileName, file);
-        }
-
-        List<ResourceDto> resources = new ArrayList<>();
-        for (String p : paths) {
-            StatObjectResponse sor = minioService.statObject(p);
-            resources.add(resourceMapper.map(sor.size(), new Path(sor.object(), userId)));
-        }
-        return resources;
-    }
-
     private void createRecursiveDirectories(Long userId, String fileName) {
-        Path tempPath = new Path(fileName, userId);
+        Path total = new Path(fileName, userId);
 
         StringBuilder sb = new StringBuilder();
-        String[] split = tempPath.getPathWithoutPrefixAndFile().split("/");
+        String[] split = total.getPathWithoutPrefixAndFile().split("/");
         for (String s : split) {
             sb.append(s).append("/");
             createDirectory(sb.toString(), userId);
         }
-    }
-
-    @Override
-    public ResourceDto createDirectory(String path, Long userId) {
-        path = path.endsWith("/") ? path : path + "/";
-        Path total = new Path(path, userId);
-        ObjectWriteResponse owr = minioService.putEmptyObject(total.getAbsolutePath());
-        return resourceMapper.map(null, new Path(owr.object(), userId));
-    }
-
-    @Override
-    public List<ResourceDto> getDirectory(String path, Long userId) {
-        Path total = new Path(path, userId);
-        Iterable<Result<Item>> results = minioService.listObjects(total.getAbsolutePath(), true);
-
-        List<ResourceDto> resources = new ArrayList<>();
-        for (Result<Item> result : results) {
-            try {
-                Item item = result.get();
-                resources.add(resourceMapper.map(item.size(), new Path(item.objectName(), userId)));
-            } catch (ErrorResponseException | InsufficientDataException | InternalException | InvalidKeyException |
-                     InvalidResponseException | IOException | NoSuchAlgorithmException | ServerException |
-                     XmlParserException e) {
-                throw new RuntimeException(e);
-            }
-        }
-        return resources;
-    }
-
-    @Override
-    public void createUserDirectory(User user) {
-        minioService.putEmptyObject(USER_PREFIX.formatted(user.getId()));
     }
 
 }
